// src/app/enums/map.enum.ts
export enum PRODUCTION_MAP_KEY_COLOUR {
  OVER_500 = '#660099',
  OVER_100 = '#CC00FF',
  OVER_50 = '#CC99FF',
  OVER_10 = '#FFFF00',
  OVER_5 = '#CCFF33',
  OVER_1 = '#669900',
  OVER_0 = '#336600',
  ZERO = '#999999',
  BLANK = '',
}

export enum PRODUCTION_MAP_KEY_MIN_VALUE {
  OVER_500 = 500.1,
  OVER_100 = 100.01,
  OVER_50 = 50.01,
  OVER_10 = 10.01,
  OVER_5 = 5.01,
  OVER_1 = 1.01,
  OVER_0 = 0.01,
  ZERO = 0,
}

export enum PRODUCTION_MAP_KEY_MAX_VALUE {
  UNDER_1000 = 1000,
  UNDER_500 = 500,
  UNDER_100 = 100,
  UNDER_50 = 50,
  UNDER_10 = 10,
  UNDER_5 = 5,
  UNDER_1 = 1,
}